requirements.txt
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
colorama==0.4.4
cryptography==37.0.2
execnet==1.9.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
MouseInfo==0.1.3
outcome==1.1.0
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
py==1.11.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.1.2
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-xdist==2.5.0
pytweening==1.0.4
selenium==4.1.5
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
